requirements.txt
bidict==0.23.1
click==8.1.7
dnspython==2.6.1
eventlet==0.36.1
Flask==2.0.3
Flask-SocketIO==5.1.1
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
python-engineio==4.3.1
python-socketio==5.5.0
setuptools==72.1.0
six==1.16.0
Werkzeug==2.0.3
httpx[http2]==0.24.0
pyjwt==2.6.0
cryptography==41.0.2